// WebUI/app-frontend/src/app/enums/match-score-type.ts
export enum MatchScoreType {
    TacticalPerformance = 0,
    BallControl = 1,
    LeftFootPassAccuracy = 2,
    RightFootPassAccuracy = 3,
    LeftFootDribblingAccuracy = 4,
    RightFootDribblingAccuracy = 5,
    LeftFootBallReceivingAccuracy = 6,
    RightFootBallReceivingAccuracy = 7,
    LeftFootShotsAccuracy = 8,
    RightFootShotsAccuracy = 9,
    HeadingAccuracy = 10,
    OneVsOneScore = 11, 
    Mobility = 12,
    Strength = 13, 
    Endurance = 14,
    Agility = 15,
    Coordination = 16,
    Concentration = 17,
    EmotionsControl = 18,
    Selfconfidence = 19,
    StressControl = 20,
    Attitude = 21,
    Communication = 22,
    Cooperation = 23,
    Determination = 24,
    Discipline = 25,
    Engagement = 26,
    Creativity = 27,
    Decisiveness = 28,
    Awareness = 29
}